requirements.txt
python-garminconnect
supabase
